# specs/openapi.yaml
openapi: 3.0.0
info:
  title: Restaurant Ordering API
  description: A restaurant ordering system with authentication and restaurant management
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.restaurant.example.com/v1
    description: Production server
security:
  - BearerAuth: []
  - GoogleOIDC:
      - openid
      - email
      - profile
paths:
...
  /orders:
    get:
      summary: List all orders
      description: Retrieves a list of orders. Can be filtered by restaurant, user, and/or status. Users can only see their own orders unless they have ADMIN or RESTAURANT_OWNER role.
      operationId: getOrders
      security:
        - BearerAuth: []
      parameters:
        - name: restaurantId
          in: query
          description: Filter orders by restaurant ID
          schema:
            type: string
            format: uuid
        - name: userId
          in: query
          description: Filter orders by user ID
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: Filter orders by status
          schema:
            type: string
            enum: [pending, preparing, ready, completed, cancelled]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
        "400":
          description: Invalid query parameters
        "401":
          description: Unauthorized - invalid or missing token
    post:
      summary: Create a new order
      description: Creates a new order for a restaurant. Customer must be authenticated. Order must contain at least one item.
      operationId: createOrder
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid request
        "401":
          description: Unauthorized - authentication required
        "500":
          description: Internal server error
  /orders/{orderId}:
    get:
      summary: Get a specific order
      description: Retrieves detailed information about a specific order by its ID. Users can only access their own orders unless they have ADMIN or RESTAURANT_OWNER role.
      operationId: getOrder
      security:
        - BearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          description: Unique identifier of the order
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid orderId format
        "401":
          description: Unauthorized - authentication required
        "404":
          description: Order not found
    put:
      summary: Update an order
      description: Updates an existing order. Only allowed for orders in 'pending' status unless user has ADMIN role. Can update items, customer info, and status.
      operationId: updateOrder
      security:
        - BearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          description: Unique identifier of the order to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderRequest"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid orderId format or request body
        "401":
          description: Unauthorized - authentication required
        "403":
          description: Forbidden - order not in pending status or user lacks permission
        "404":
          description: Order not found
        "500":
          description: Internal server error
    delete:
      summary: Delete an order
      description: Cancels and deletes an order. Only the order owner or users with ADMIN/RESTAURANT_OWNER role can delete orders.
      operationId: deleteOrder
      security:
        - BearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          description: Unique identifier of the order to delete
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Order deleted
        "400":
          description: Invalid orderId format
        "401":
          description: Unauthorized - authentication required
        "404":
          description: Order not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for Spring Security authentication
    GoogleOIDC:
      type: openIdConnect
      description: Google OpenID Connect authentication
      openIdConnectUrl: https://accounts.google.com/.well-known/openid-configuration
  schemas:
...
    Order:
      type: object
      description: Customer order information
      required:
        - id
        - restaurantId
        - customerName
        - customerEmail
        - items
        - status
        - totalAmount
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique order identifier (UUID)
        restaurantId:
          type: string
          format: uuid
          description: ID of the restaurant fulfilling the order
        customerId:
          type: string
          format: uuid
          nullable: true
          description: User ID who placed the order (null for guest orders)
        customerName:
          type: string
          description: Customer name (for guest orders or display purposes)
        customerEmail:
          type: string
          description: Customer email for order notifications
        items:
          type: array
          description: List of items in the order
          items:
            $ref: "#/components/schemas/OrderItem"
        status:
          type: string
          enum: [pending, preparing, ready, completed, cancelled]
          description: Current order status
        totalAmount:
          type: number
          description: Total order amount in dollars
        specialInstructions:
          type:
            - "null"
            - string
          description: Special requests or dietary requirements
        createdAt:
          type: string
          format: date-time
          description: Order creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last status update timestamp
    OrderItem:
      type: object
      description: Individual item within an order
      properties:
        itemId:
          type: string
          description: Menu item identifier
        name:
          type: string
          description: Menu item name
        quantity:
          type: integer
          minimum: 1
          description: Quantity ordered
        unitPrice:
          type: number
          description: Price per unit in dollars
        subtotal:
          type: number
          description: Total for this item (quantity * unitPrice) in dollars
    CreateOrderRequest:
      type: object
      description: Request body for creating a new order
      additionalProperties: false
      required:
        - restaurantId
        - customerName
        - customerEmail
        - items
      properties:
        restaurantId:
          type: string
          format: uuid
          description: Restaurant ID to place order with
        customerName:
          type: string
          description: Name of the customer
          minLength: 1
          maxLength: 100
          pattern: "^\\S.*\\S$|^\\S$"
        customerEmail:
          type: string
          description: Email for order notifications
          pattern: "^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$"
        items:
          type: array
          description: List of items to order
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required:
              - itemId
              - quantity
            properties:
              itemId:
                type: string
                description: ID of the menu item
              quantity:
                type: integer
                description: Number of this item to order
                minimum: 1
        specialInstructions:
          type: string
          description: Special requests or dietary requirements (optional)
          maxLength: 500
    UpdateOrderRequest:
      type: object
      description: Request body for updating an existing order
      additionalProperties: false
      properties:
        customerName:
          type: string
          description: Name of the customer (optional)
          minLength: 1
          maxLength: 100
        customerEmail:
          type: string
          format: email
          description: Email for order notifications (optional)
        items:
          type: array
          description: Updated list of items (optional)
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required:
              - itemId
              - quantity
            properties:
              itemId:
                type: string
                description: ID of the menu item
              quantity:
                type: integer
                description: Number of this item to order
                minimum: 1
        status:
          type: string
          description: New order status (optional, restricted based on current status)
          enum: [pending, preparing, ready, completed, cancelled]
        specialInstructions:
          type: string
          description: Special requests or dietary requirements (optional)
          maxLength: 500
...